Update 1-2-proj-arch.qmd
Suggesting a correction of an index entry for the term "Google Sheets" in the output of a Quarto document - currently the Google Sheet name does not generate resulting in a following sentence:

"If you’re still early in your project lifecycle, a can be a great way to save and recall a flat file. "

correction:

"If you’re still early in your project lifecycle, a "Google Sheet" can be a great way to save and recall a flat file. "